+++ MiniApp/Classes/Extensions/MiniApp+WKUserContentController.swift
@@ -6,8 +6,8 @@ extension WKUserContentController {
     /// - Parameters:
     ///   - delegate: Callback protocol to handle the message from the WebView
     ///   - hostMessageInterface: Message interface protocol of the host application
-    func addMiniAppScriptMessageHandler(delegate: MiniAppCallbackProtocol, hostAppMessageDelegate: MiniAppMessageProtocol) {
-        add(MiniAppScriptMessageHandler(delegate: delegate, hostAppMessageDelegate: hostAppMessageDelegate), name: Constants.javascriptInterfaceName)
+    func addMiniAppScriptMessageHandler(delegate: MiniAppCallbackProtocol, hostAppMessageDelegate: MiniAppMessageProtocol, miniAppId: String) {
+        add(MiniAppScriptMessageHandler(delegate: delegate, hostAppMessageDelegate: hostAppMessageDelegate, miniAppId: miniAppId), name: Constants.javascriptInterfaceName)
     }
 
     /// Method to remove the custom WKScriptMessageHandler. Removing the message handler will help to avoid memory leaks