+++ MiniApp/Classes/Models/MiniAppJavaScriptMessageInfo.swift
@@ -6,9 +6,23 @@ struct MiniAppJavaScriptMessageInfo: Decodable {
 
 struct RequestParameters: Decodable {
     let permission: String?
+    let permissions: [MiniAppCustomPermissionsRequest]?
     let locationOptions: LocationOptions?
 }
 
 struct LocationOptions: Decodable {
     let enableHighAccuracy: Bool?
+}
+
+struct MiniAppCustomPermissionsRequest: Decodable {
+    let name: String?
+    let description: String?
+}
+
+struct MiniAppCustomPermissionsResponse: Codable {
+    let permissions: [MiniAppCustomPermissionsListResponse]
+}
+
+struct MiniAppCustomPermissionsListResponse: Codable {
+    let name, isGranted: String
 }